requirements.txt
google-generativeai==0.8.3
python-dotenv
fastapi
uvicorn[standard]
